+++ src/globals.php
@@ -9,7 +9,6 @@ declare(strict_types=1);
 
 namespace OceanMoon\Core;
 
-use DateTimeInterface;
 use ErrorException;
 use Throwable;
 
@@ -37,8 +36,6 @@ const RECURSION = '*RECURSION*';
  * This can produce a notice or warning for some values (arrays, closures, objects that are not Stringable).
  *
  * The function name mimics Java, Scala, Swift, Rust, Go, Julia, etc., and aligns with PHP's print() construct.
- *
- * Provided here for completeness, but writeln() is generally better.
  *
  * @param mixed $value [optional] The value to print.
  */
@@ -94,16 +91,15 @@ function ex(mixed $value): string
  * Convert any value to a string, without errors.
  *
  * - Default string conversion is used when it works.
- * - Special handling for datetimes, which don't provide a default conversion to string.
- * - Falls back to stringify() for values that emit warnings or errors on default string conversion.
+ * - Falls back to stringify() for values that emit warnings or errors on default string conversion (e.g. arrays, NAN, \
+ *   non-Stringable objects).
  *
  * @param mixed $value Whatever you want converted to a string.
  * @return string The value as a string.
  */
 function to_string(mixed $value): string
 {
-    // Temporarily convert warnings to exceptions, so we can catch cases where the default cast to string would
-    // emit a warning (e.g. arrays, NAN).
+    // Temporarily convert warnings to exceptions to catch cases where the default cast would emit a warning.
     set_error_handler(static fn () => throw new ErrorException());
 
     try {
@@ -112,11 +108,6 @@ function to_string(mixed $value): string
         // Fall through to the DateTimeInterface/Stringify handling below.
     } finally {
         restore_error_handler();
-    }
-
-    // Special handling for datetimes.
-    if ($value instanceof DateTimeInterface) {
-        return $value->format(DateTimeInterface::ATOM);
     }
 
     // Fallback to stringify() which will handle anything else.